backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
redis>=4.5.4,<5.0.0
python-dotenv>=0.19.0,<0.20.0
websockets>=10.0,<11.0
pydantic>=1.8.0,<2.0.0
requests>=2.26.0,<3.0.0
typing-extensions>=4.0.0,<5.0.0
aioredis>=2.0.0,<3.0.0
